[mod] rename block view to 'blockCount' to avoid confusion with other 'Members' views.
