+++ views/MENU.html
@@ -3,7 +3,7 @@
     <td><a href="willMembers">willMembers</a></td>
     <td><a href="members">members</a></td>
     <td><a href="membersCount">membersCount</a></td>
-    <td><a href="blockMembers">blockMembers</a></td>
+    <td><a href="blockCount">blockCount</a></td>
     <td><a href="monetaryMass">monetaryMass</a></td>
     <td><a href="about">about</a></td>
   </tr>
